# docs/openapi.yaml
openapi: 3.1.0
info:
  title: HomeDeck API
  version: 1.0.0
  description: |
    Все маршруты `/api/v1/*`, кроме setup, login, session и public, требуют сессии.
    Изменяющие запросы с сессией требуют заголовок `X-CSRF-Token` из `GET /api/v1/session`.
    Обновление сущностей с ревизией требует `If-Match: "<revision>"`; несовпадение возвращает 409.
    Время — RFC 3339 в UTC; длительности — целые секунды; ID — строки с префиксом типа.
servers:
  - url: /
...
paths:
...
  /api/v1/setup:
    get:
      tags: [auth]
      summary: Нужна ли первичная настройка
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
                required: [required]
                properties:
                  required: { type: boolean }
    post:
      tags: [auth]
      summary: Создать администратора по одноразовому setup-token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token, username, password, title]
              properties:
                token: { type: string }
                username: { type: string, minLength: 1, maxLength: 64 }
                password: { type: string, minLength: 8, maxLength: 256 }
                title: { type: string, maxLength: 100 }
      responses:
        "200":
          description: Администратор создан, сессия установлена cookie
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Session" }
        "403": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
        "429": { $ref: "#/components/responses/Error" }
  /api/v1/login:
    post:
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username: { type: string }
                password: { type: string }
      responses:
        "200":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Session" }
        "401": { $ref: "#/components/responses/Error" }
        "429": { $ref: "#/components/responses/Error" }
...
  /api/v1/session:
    get:
      tags: [auth]
      responses:
        "200":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Session" }
        "401": { $ref: "#/components/responses/Error" }
  /api/v1/password:
    put:
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [current, next]
              properties:
                current: { type: string }
                next: { type: string, minLength: 8 }
      responses:
        "204": { description: ok }
  /api/v1/settings:
    get:
      tags: [settings]
      responses:
        "200":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Settings" }
    put:
      tags: [settings]
      parameters: [{ $ref: "#/components/parameters/IfMatch" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/Settings" }
      responses:
        "200":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Settings" }
  /api/v1/pages:
    get:
      tags: [pages]
      summary: Список страниц без виджетов
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/PageSummary" }
    post:
      tags: [pages]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PageInput" }
      responses:
        "201":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Page" }
  /api/v1/pages/{id}:
    parameters: [{ $ref: "#/components/parameters/Id" }]
    get:
      tags: [pages]
      summary: Страница целиком; id или slug
      responses:
        "200":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Page" }
    put:
      tags: [pages]
      summary: Атомарное сохранение страницы вместе с группами и виджетами
      parameters: [{ $ref: "#/components/parameters/IfMatch" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PageInput" }
      responses:
        "200":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Page" }
        "409": { $ref: "#/components/responses/Error" }
    delete:
      tags: [pages]
      responses:
        "204": { description: ok }
  /api/v1/services:
    get:
      tags: [catalog]
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Service" }
    post:
      tags: [catalog]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ServiceInput" }
      responses:
        "201":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Service" }
  /api/v1/services/{id}:
    parameters: [{ $ref: "#/components/parameters/Id" }]
    get:
      tags: [catalog]
      responses:
        "200":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Service" }
    put:
      tags: [catalog]
      parameters: [{ $ref: "#/components/parameters/IfMatch" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ServiceInput" }
      responses:
        "200":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Service" }
    delete:
      tags: [catalog]
      responses:
        "204": { description: ok }
  /api/v1/checks:
    get:
      tags: [monitoring]
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Check" }
    post:
      tags: [monitoring]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CheckInput" }
      responses:
        "201":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Check" }
  /api/v1/checks/{id}:
    parameters: [{ $ref: "#/components/parameters/Id" }]
    get:
      tags: [monitoring]
      responses:
        "200":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Check" }
    put:
      tags: [monitoring]
      parameters: [{ $ref: "#/components/parameters/IfMatch" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CheckInput" }
      responses:
        "200":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Check" }
    delete:
      tags: [monitoring]
      responses:
        "204": { description: ok }
  /api/v1/sources:
    get:
      tags: [monitoring]
      summary: Пользовательские и системные источники
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Source" }
    post:
      tags: [monitoring]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SourceInput" }
      responses:
        "201":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Source" }
  /api/v1/sources/test:
    post:
      tags: [monitoring]
      summary: Проверка несохранённого источника
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SourceInput" }
      responses:
        "200":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SourceTestResult" }
  /api/v1/sources/{id}:
    parameters: [{ $ref: "#/components/parameters/Id" }]
    get:
      tags: [monitoring]
      responses:
        "200":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Source" }
    put:
      tags: [monitoring]
      parameters: [{ $ref: "#/components/parameters/IfMatch" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SourceInput" }
      responses:
        "200":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Source" }
        "403": { $ref: "#/components/responses/Error" }
    delete:
      tags: [monitoring]
      responses:
        "204": { description: ok }
  /api/v1/sources/{id}/test:
    parameters: [{ $ref: "#/components/parameters/Id" }]
    post:
      tags: [monitoring]
      summary: Ограниченная проверка соединения без изменения настроек
      responses:
        "200":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SourceTestResult" }
  /api/v1/devices:
    get:
      tags: [monitoring]
      summary: Устройства с текущим состоянием и последними значениями
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Device" }
    post:
      tags: [monitoring]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/DeviceInput" }
      responses:
        "201":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Device" }
        "422": { $ref: "#/components/responses/Error" }
  /api/v1/devices/test:
    post:
      tags: [monitoring]
      summary: Однократный опрос несохранённого устройства
      description: Ключ или учётные данные можно передать в поле secret вместо secret_id — так новое устройство проверяется до сохранения ключа.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - $ref: "#/components/schemas/DeviceInput"
                - type: object
                  properties:
                    secret: { $ref: "#/components/schemas/SecretInput" }
      responses:
        "200":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DeviceStatus" }
        "422": { $ref: "#/components/responses/Error" }
  /api/v1/devices/{id}:
    parameters: [{ $ref: "#/components/parameters/Id" }]
    get:
      tags: [monitoring]
      responses:
        "200":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Device" }
    put:
      tags: [monitoring]
      parameters: [{ $ref: "#/components/parameters/IfMatch" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/DeviceInput" }
      responses:
        "200":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Device" }
        "409": { $ref: "#/components/responses/Error" }
    delete:
      tags: [monitoring]
      responses:
        "204": { description: ok }
  /api/v1/secrets:
    get:
      tags: [monitoring]
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Secret" }
    post:
      tags: [monitoring]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SecretInput" }
      responses:
        "201":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Secret" }
  /api/v1/secrets/{id}:
    parameters: [{ $ref: "#/components/parameters/Id" }]
    put:
      tags: [monitoring]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SecretInput" }
      responses:
        "200":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Secret" }
    delete:
      tags: [monitoring]
      responses:
        "204": { description: ok }
        "409": { $ref: "#/components/responses/Error" }
  /api/v1/presets:
    get:
      tags: [monitoring]
      summary: Встроенные шаблоны и сохранённые запросы
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/QueryPreset" }
    post:
      tags: [monitoring]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/QueryPresetInput" }
      responses:
        "201":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/QueryPreset" }
  /api/v1/presets/{id}:
    parameters: [{ $ref: "#/components/parameters/Id" }]
    put:
      tags: [monitoring]
      parameters: [{ $ref: "#/components/parameters/IfMatch" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/QueryPresetInput" }
      responses:
        "200":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/QueryPreset" }
    delete:
      tags: [monitoring]
      responses:
        "204": { description: ok }
...
  /api/v1/assets:
    get:
      tags: [catalog]
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Asset" }
    post:
      tags: [catalog]
      summary: PNG, JPEG или WebP до 5 MiB
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file: { type: string, format: binary }
      responses:
        "201":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Asset" }
        "413": { $ref: "#/components/responses/Error" }
        "415": { $ref: "#/components/responses/Error" }
  /api/v1/assets/{id}:
    parameters: [{ $ref: "#/components/parameters/Id" }]
    delete:
      tags: [catalog]
      responses:
        "204": { description: ok }
        "409": { $ref: "#/components/responses/Error" }
  /assets/{id}:
    parameters: [{ $ref: "#/components/parameters/Id" }]
    get:
      tags: [catalog]
      summary: Содержимое загруженного файла, без авторизации
      responses:
        "200": { description: file }
  /api/v1/import/preview:
    post:
      tags: [portability]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [format, file]
              properties:
                format: { type: string, enum: [homedeck, homer] }
                file: { type: string, format: binary, description: YAML }
                assets: { type: string, format: binary, description: ZIP с иконками Homer, пути относительно конфига }
      responses:
        "200":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ImportPreview" }
        "422": { $ref: "#/components/responses/Error" }
  /api/v1/import/apply:
    post:
      tags: [portability]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token]
              properties:
                token: { type: string }
      responses:
        "200":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ImportPreview" }
        "409": { $ref: "#/components/responses/Error" }
...
  /api/v1/public:
    get:
      tags: [public]
      summary: Опубликованная страница; 404, если публикация выключена
      responses:
        "200":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PublicPage" }
        "404": { $ref: "#/components/responses/Error" }
  /api/v1/public/widgets/{id}/data:
    parameters:
      - { $ref: "#/components/parameters/Id" }
      - name: range
        in: query
        schema: { $ref: "#/components/schemas/RangeName" }
      - name: points
        in: query
        schema: { type: integer, minimum: 10, maximum: 1000 }
    get:
      tags: [public]
      summary: Данные опубликованного виджета; выражение выбирает сервер
      responses:
        "200":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/WidgetData" }
  /api/v1/widgets/{id}/data:
    parameters:
      - { $ref: "#/components/parameters/Id" }
      - name: range
        in: query
        schema: { $ref: "#/components/schemas/RangeName" }
      - name: start
        in: query
        schema: { type: string, format: date-time }
      - name: end
        in: query
        schema: { type: string, format: date-time }
      - name: points
        in: query
        schema: { type: integer, minimum: 10, maximum: 1000 }
    get:
      tags: [pages]
      summary: Данные сохранённого виджета number, chart, status или link
      responses:
        "200":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/WidgetData" }
components:
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema: { type: string }
    IfMatch:
      name: If-Match
      in: header
      required: true
      schema: { type: string, example: '"3"' }
  responses:
    Error:
      description: Ошибка
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
...
    Session:
      type: object
      required: [user, csrf_token]
      properties:
        user:
          type: object
          properties:
            id: { type: string }
            username: { type: string }
        csrf_token: { type: string }
    Settings:
      type: object
      properties:
        title: { type: string }
        logo_asset_id: { type: [string, "null"] }
        start_page_id: { type: [string, "null"] }
        public_page_id: { type: [string, "null"], description: null — публичная страница выключена }
        revision: { type: integer, readOnly: true }
        restart_required:
          type: array
          readOnly: true
          items: { type: string }
          description: Параметры запуска, отличающиеся от сохранённых ожиданий
        runtime:
          type: object
          readOnly: true
          properties:
            listen: { type: string }
            data_dir: { type: string }
            retention: { type: string }
            version: { type: string }
            tsdb_version: { type: string }
...
    PageSummary:
      type: object
      properties:
        id: { type: string }
        title: { type: string }
        slug: { type: string }
        order: { type: integer }
        revision: { type: integer }
...
    RangeName:
      type: string
      enum: [1h, 6h, 24h, 7d, 30d]
...